deep_search_labeling/backend/requirements.txt
# Web Framework
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Data Processing
pandas>=2.0.0
numpy>=1.21.0
datasets>=2.12.0

# Export Formats
spacy>=3.7.0,<3.8.0
transformers>=4.30.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
aiofiles>=23.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# CORS and Security
fastapi-cors>=0.0.6